Opa and opb are both quantized from 4 banks, and results are writen into 4 banks
