+++ apps/llama2c260K_q_vifmm/main.c
@@ -399,49 +399,49 @@ float vifbw_e32_m4( float* x, int8_t* w, int size) {
     float* x_   = x;
     float  sum;
     unsigned long int block_size_max=65535;
-
+  
     asm volatile("vsetvli %0, %1, e32, m4, ta, ma" : "=r"(block_size) : "r"(block_size_max));
     asm volatile("vmv.v.i v8,  0");
     asm volatile("vmv.v.i v0,  0");
-
+  
     // printf("Available block_size=%d\n", block_size);
-
+  
     if (size < block_size)
     {
+        asm volatile("vle32.v v16, (%0);" ::"r"(x_));
         asm volatile("vsetvli zero, %0, e8, m1, ta, ma" ::"r"(size));    
         asm volatile("vle8.v v24, (%0);" ::"r"(w_));
         asm volatile("vsetvli zero, %0, e32, m4, ta, ma" ::"r"(size));    
-        asm volatile("vle32.v v16, (%0);" ::"r"(x_));
       #ifndef LLVM
-        asm volatile(".word 0xbb882057");   // gcc  vifbw v0, v24, v16
+        asm volatile(".word 0xbb882057");   // gcc  vifbw v0, v16, v24
       #else
-      //   asm volatile(".word 0xbb0c2057");    // llvm vifbw.vv v0, v16, v24
+      //   asm volatile(".word 0xbb882057");    // llvm vifbw.vv v0, v16, v24
         asm volatile("vifbw.vv v0, v16, v24");
       #endif
-
+  
         
     } else {
     
         for (unsigned long int m = 0; m < size; m += block_size) {
-          const unsigned long int p_ = MIN(size - m, block_size);
+          const unsigned long int p_ = MIN(size - m, block_size); 
+          asm volatile("vle32.v v16, (%0);" ::"r"(x_));
           asm volatile("vsetvli zero, %0, e8, m1, ta, ma" ::"r"(p_));    
           asm volatile("vle8.v v24, (%0);" ::"r"(w_));
-          asm volatile("vsetvli zero, %0, e32, m4, ta, ma" ::"r"(p_));    
-          asm volatile("vle32.v v16, (%0);" ::"r"(x_));
-        #ifndef LLVM
-          asm volatile(".word 0xbb882057");   // gcc  vifbw v0, v24, v16
-        #else
-        //   asm volatile(".word 0xbb0c2057");    // llvm vifbw.vv v0, v16, v24
-          asm volatile("vifbw.vv v0, v16, v24");
-        #endif
           w_ += block_size;
           x_ += block_size;
+          asm volatile("vsetvli zero, %0, e32, m4, ta, ma" ::"r"(p_));    
+        #ifndef LLVM
+          asm volatile(".word 0xbb882057");   // gcc  vifbw v0, v16, v24
+        #else
+        //   asm volatile(".word 0xbb882057");    // llvm vifbw.vv v0, v16, v24
+          asm volatile("vifbw.vv v0, v16, v24");
+        #endif
         }
     }
     asm volatile("vfredsum.vs v8, v0, v8");  // vredsum.vs vd, vs2,vs1; vd[0]=sum(vs1[0], vs2[*])
     asm volatile("vfmv.f.s %0, v8;":"=f"(sum));
     return sum;
-}
+  }
 
 void matmul_vifmm(float* xout, float* x, QuantizedTensor *w, int n, int d) {
     // W (d,n) @ x (n,) -> xout (d,)
