+++ hardware/ifmix_src/operand_queue_mfpu_a.sv
@@ -32,6 +32,7 @@ module operand_queue_mfpu_a import ara_pkg::*; import rvv_pkg::*; import cf_math
     output logic                              cmd_pop_o,
     // Interface with the Vector Register File
     input  elen_t                             operand_i,
+    input  elen_t  [2:0]                      operand_vifmm_i,
     input  logic                              operand_valid_i,
     input  logic                              operand_issued_i,
     output logic                              operand_queue_ready_o,
@@ -120,14 +121,27 @@ module operand_queue_mfpu_a import ara_pkg::*; import rvv_pkg::*; import cf_math
     end
   end
 
+  elen_t  [3:0]     operand_fifo;
+
   always_comb begin
-    operand_push_valid_d = {operand_push_valid[2:0], operand_push_valid[3]};  // 1 left-shift 
+    
     ibuf_empty           = &ibuf_empty_pop[3:0];
 
-    if (transfer_all_quantize_en_i) begin
-      operand_pop_valid_d = {4{ibuf_pop}};
+    if (conv_vifmm_o == OpQueueConversionF32I8) begin
+      operand_push_valid_d = {4{operand_valid_i}};  
+      operand_fifo[0]    = operand_i;
+      operand_fifo[1]    = operand_vifmm_i[0];
+      operand_fifo[2]    = operand_vifmm_i[1];
+      operand_fifo[3]    = operand_vifmm_i[2];
+
+
+      operand_pop_valid_d = transfer_all_quantize_en_i ? {4{ibuf_pop }} : {operand_pop_valid[2:0], operand_pop_valid[3]} ;
       ibuf_operand        = ibuf_operand_pop[0];
     end else begin
+      operand_push_valid_d = {operand_push_valid[2:0], operand_push_valid[3]};  // 1 left-shift 
+      operand_fifo        = {4{operand_i}};
+
+
       operand_pop_valid_d  = {operand_pop_valid[2:0], operand_pop_valid[3]};  // 1 left-shift 
 
       ibuf_operand = {{64{operand_pop_valid[0]}} & ibuf_operand_pop[0]} | 
@@ -146,11 +160,11 @@ module operand_queue_mfpu_a import ara_pkg::*; import rvv_pkg::*; import cf_math
     .rst_ni    (rst_ni         ),
     .testmode_i(1'b0           ),
     .flush_i   (flush_i        ),
-    .data_i    (operand_i      ),
-    .push_i    (operand_push_valid[0] & operand_valid_i),
+    .data_i    (operand_fifo[0]      ),
+    .push_i    (operand_push_valid_d[0] & operand_valid_i),
     .full_o    (/* Unused */   ),
     .data_o    (ibuf_operand_pop[0]   ),
-    .pop_i     (operand_pop_valid[0] & ibuf_pop       ),
+    .pop_i     (operand_pop_valid_d[0] & ibuf_pop       ),
     .empty_o   (ibuf_empty_pop[0]     ),
     .usage_o   (/* Unused */   )
   );
@@ -164,11 +178,11 @@ module operand_queue_mfpu_a import ara_pkg::*; import rvv_pkg::*; import cf_math
     .rst_ni    (rst_ni         ),
     .testmode_i(1'b0           ),
     .flush_i   (flush_i        ),
-    .data_i    (operand_i      ),
-    .push_i    (operand_push_valid[1] & operand_valid_i),
+    .data_i    (operand_fifo[1]      ),
+    .push_i    (operand_push_valid_d[1] & operand_valid_i),
     .full_o    (/* Unused */   ),
     .data_o    (ibuf_operand_pop[1]   ),
-    .pop_i     (operand_pop_valid[1] & ibuf_pop       ),
+    .pop_i     (operand_pop_valid_d[1] & ibuf_pop       ),
     .empty_o   (ibuf_empty_pop[1]     ),
     .usage_o   (/* Unused */   )
   );
@@ -182,11 +196,11 @@ module operand_queue_mfpu_a import ara_pkg::*; import rvv_pkg::*; import cf_math
     .rst_ni    (rst_ni         ),
     .testmode_i(1'b0           ),
     .flush_i   (flush_i        ),
-    .data_i    (operand_i      ),
-    .push_i    (operand_push_valid[2] & operand_valid_i),
+    .data_i    (operand_fifo[2]      ),
+    .push_i    (operand_push_valid_d[2] & operand_valid_i),
     .full_o    (/* Unused */   ),
     .data_o    (ibuf_operand_pop[2]   ),
-    .pop_i     (operand_pop_valid[2] & ibuf_pop       ),
+    .pop_i     (operand_pop_valid_d[2] & ibuf_pop       ),
     .empty_o   (ibuf_empty_pop[2]     ),
     .usage_o   (/* Unused */   )
   );
@@ -200,11 +214,11 @@ module operand_queue_mfpu_a import ara_pkg::*; import rvv_pkg::*; import cf_math
     .rst_ni    (rst_ni         ),
     .testmode_i(1'b0           ),
     .flush_i   (flush_i        ),
-    .data_i    (operand_i      ),
-    .push_i    (operand_push_valid[3] & operand_valid_i),
+    .data_i    (operand_fifo[3]      ),
+    .push_i    (operand_push_valid_d[3] & operand_valid_i),
     .full_o    (/* Unused */   ),
     .data_o    (ibuf_operand_pop[3]   ),
-    .pop_i     (operand_pop_valid[3] & ibuf_pop       ),
+    .pop_i     (operand_pop_valid_d[3] & ibuf_pop       ),
     .empty_o   (ibuf_empty_pop[3]     ),
     .usage_o   (/* Unused */   )
   );
@@ -232,7 +246,7 @@ module operand_queue_mfpu_a import ara_pkg::*; import rvv_pkg::*; import cf_math
     if (flush_i) ibuf_usage_d = '0;
 
     // Are we ready?
-    operand_queue_ready_o = (ibuf_usage_q != DataBufDepth);
+    operand_queue_ready_o = (ibuf_usage_q != DataBufDepth/4);  // gukai@20250818
   end
 
   always_ff @(posedge clk_i or negedge rst_ni) begin: p_ibuf_usage_ff
@@ -628,6 +642,8 @@ module operand_queue_mfpu_a import ara_pkg::*; import rvv_pkg::*; import cf_math
           if (SupportIntExt8) elem_count_d = elem_count_q + (1 << (unsigned'(EW64) - unsigned'(cmd.eew))) / 8;
         OpQueueReductionZExt:
           elem_count_d = elem_count_q + 1;
+        OpQueueConversionF32I8: // gukai@20250813
+          elem_count_d = transfer_all_quantize_en_i ? (elem_count_q + 8) :(elem_count_q + (1 << (unsigned'(EW64) - unsigned'(cmd.eew))));
         default: elem_count_d = elem_count_q + (1 << (unsigned'(EW64) - unsigned'(cmd.eew)));
       endcase
